feat: handle use-generator composable
