+++ ui/src/components/grid-generator/QasGridGenerator.vue
@@ -82,9 +82,7 @@ import QasGridItem from '../grid-item/QasGridItem.vue'
 import QasHeader from '../header/QasHeader.vue'
 
 import useGenerator, { baseProps } from '../../composables/private/use-generator'
-import { isEmpty, humanize, filterObject } from '../../helpers'
 import { useScreen } from '../../composables'
-import isObject from 'lodash-es/isObject'
 import { computed } from 'vue'
 
 // define component name
@@ -156,7 +154,13 @@ const props = defineProps({
 })
 
 // composables
-const { classes, getFieldClass } = useGenerator({ props, isGrid: true })
+const {
+  classes,
+  getFieldClass,
+  getHeaderProps,
+  getNormalizedFields,
+  getFieldsByResult
+} = useGenerator({ props, isGrid: true })
 
 // computeds
 const component = computed(() => {
@@ -177,9 +181,8 @@ const headerClass = computed(() => {
 })
 
 const normalizedFields = computed(() => {
-  /**
-   * No caso de não ter "fieldset", virá via props o result e fields.
-   */
+  console.log('to linkado')
+  // No caso de não ter "fieldset", virá via props o result e fields.
   if (!hasFieldset.value) {
     return {
       default: {
@@ -190,100 +193,18 @@ const normalizedFields = computed(() => {
     }
   }
 
-  /**
-   * No caso de ter "fieldset", deve organizar os fields/result de acordo com o fieldset.
-   */
-  return getNormalizedFields(props.fieldset)
+  // No caso de ter "fieldset", deve normalizar os fields/result de acordo com o fieldset.
+  return getNormalizedFields({
+    items: props.fieldset,
+    fields: props.fields,
+    result: props.result,
+    isGrid: true
+  })
 })
 
 const hasFieldset = computed(() => !!Object.keys(props.fieldset).length)
 
 // functions
-function getFieldsByResult ({ fields, result }) {
-  const hasFields = Object.keys(fields).length
-  const hasResult = Object.keys(result).length
-
-  if (!hasResult || !hasFields) return {}
-
-  const unformattedResult = { ...result }
-  const fieldsByResult = {}
-
-  const formattedFields = getFormattedFields({ fields, result, hasResult })
-
-  for (const key in formattedFields) {
-    const field = formattedFields[key] || {}
-
-    if (!field.type) continue
-
-    const humanizedResult = humanize(field, unformattedResult[key])
-    const formattedResult = isEmpty({ value: humanizedResult }) ? props.emptyResultText : humanizedResult
-
-    fieldsByResult[key] = {
-      ...field,
-      formattedResult
-    }
-  }
-
-  return fieldsByResult
-}
-
-// TODO: Adicionar no use-generator
-function getNormalizedFields (items) {
-  const result = {}
-
-  for (const key in items) {
-    const item = items[key]
-
-    // Propriedades que são recebidas em cada fieldset / subset desconstruida
-    const {
-      label,
-      description,
-      headerProps,
-      column,
-      buttonProps,
-      fields = [],
-      subset = {},
-      useSeparator
-    } = item
-
-    // Valida caso tenha um header
-    const hasHeader = !!(label || description || Object.keys(headerProps || {}).length)
-
-    // Valida caso tenha um subset dentro de um fieldset
-    const hasSubset = !!Object.keys(subset).length
-
-    // Estrutura de um fieldset e subset
-    const structure = {
-      label,
-      description,
-      column,
-      buttonProps,
-      headerProps,
-      useSeparator,
-      fields: {},
-      subset: hasSubset ? getNormalizedFields(subset) : {},
-
-      // Propriedades auxiliares para controle na view
-      __hasHeader: hasHeader,
-      __hasSubset: hasSubset,
-      __hasFieldset: true
-    }
-
-    // Pegar os fields com base na key.
-    const fieldsByfieldsKeys = filterObject(props.fields, fields)
-
-    // Adicionar "formattedResult" com base no result.
-    const formattedFields = getFieldsByResult({ fields: fieldsByfieldsKeys, result: props.result })
-
-    // Adiciona o fields no structure
-    Object.assign(structure.fields, { ...formattedFields })
-
-    result[key] = structure
-  }
-
-  return result
-}
-
 function getContainerClasses ({ key }) {
   if (props.useInline) return 'row justify-between col-12'
 
@@ -320,32 +241,6 @@ function getGridItemProps (field) {
   }
 }
 
-/**
- * Retorna os "fields" formatados, caso a propriedade "useEmptyResult" seja "true", retorna todos os "fields", mesmo que não tenha resultado.
- * Caso a propriedade "useEmptyResult" seja "false", retorna apenas os "fields" que possuem resultado.
-*/
-function getFormattedFields ({ fields, result, hasResult }) {
-  if (props.useEmptyResult) return fields
-
-  if (!hasResult) return {}
-
-  const formattedFields = {}
-
-  for (const key in fields) {
-    const currentResult = result[key]
-
-    const validate = Array.isArray(currentResult)
-      ? currentResult.length
-      : isObject(currentResult) ? Object.keys(currentResult).length : result
-
-    if (validate) {
-      formattedFields[key] = fields[key]
-    }
-  }
-
-  return formattedFields
-}
-
 function hasSeparator ({ item, isSubset }) {
   /**
    * No caso de for subset, por padrão terá o separator, à não ser que for passado
@@ -357,26 +252,5 @@ function hasSeparator ({ item, isSubset }) {
 
   // Caso seja um fieldset, verificar se foi passado a chave "useSeparator".
   return item.useSeparator
-}
-
-// TODO: Adicionar no use-generator
-function getHeaderProps ({ values, isSubset }) {
-  const typography = isSubset ? 'h5' : 'h4'
-
-  // Separa as propriedades de label e o restante.
-  const { labelProps, ...headerProps } = values.headerProps || {}
-
-  return {
-    description: values.description,
-
-    // Mesmo que passado a tipografia no headerProps, a tipografia é sobreposta para manter a hierarquia sempre.
-    labelProps: {
-      ...labelProps,
-      ...(values.label && { label: values.label }),
-      typography
-    },
-
-    ...headerProps
-  }
 }
 </script>
